requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.1
paho-mqtt==1.4.0
pyparsing==2.3.1
python-dateutil==2.8.0
six==1.12.0
